Move keyboard renderer
